+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 ## Unreleased
 ## Release
+### 4.3.0
+- Remove legacy code
+
+#### Update webpack.config.js
+- Update `process.env` format for `webpack.DefinePlugin`
+- Update `resolve.modules` to retrieve modules from the `node_modules` of this repo firstly
+
+#### Adapt new packages
+- Add new dependencies `@twreporter/core` and `@twreporter/index-page`
+- Move some utils and constants to new package `@twreporter/core`
+- Move `BookmarkList` and `BookmarkWidget` to `@twreporter/react-components` 
+- Retrive index components from `@twreporter/index-page`
+
+#### Update .circleci/config.yml
+- Use `yarn.lock` as the indicator of the change of dependencies
+- Run build only on certain branches
+- Remove test
+
+#### Move Redux actions, reducers, configureStore to @twreporter/redux
+- Remove actions, reducers, configureStore in this repo
+- Bump `@twreporter/redux` version to 5.0.1
+
+#### Update about-us page
+- Update about-us page in include awards change
+- Update about-us page to include staffing change
+
+#### Bug Fix
+- Bug fix: Use `ogImage` rather than `leadingImage` as og image on topic landing page
+
 ### 4.2.1 
 #### Dependency Update 
 - @twreporter/react-components@^6.1.4